+++ src/compiler/scala/tools/nsc/doc/ModelExtractor.scala
@@ -113,8 +113,8 @@ trait ModelExtractor {
           sym.privateWithin.nameString
         else null
       )
-      def f(flag: Int, str: String) {
-        if (sym hasFlag flag.toLong) string = string + " " + str
+      def f(flag: Long, str: String) {
+        if (sym hasFlag flag) string = string + " " + str
       }
       f(Flags.IMPLICIT, "implicit")
       f(Flags.SEALED, "sealed")
